Move version detection into own file
