+++ bench/version.rs
@@ -1,0 +1,1 @@
+../version.rs
